update_style_responsive
style, responsive and image
