--- web-dien-lanh/src/components/content/maintenance/ACC/ACC.scss
@@ -1,14 +1,0 @@
-.your-table-class {
-    border-collapse: collapse;
-    width: 100%;
-  }
-  
-  .your-table-class th, .your-table-class td {
-    border: 1px solid #dddddd;
-    text-align: center;
-    padding: 8px;
-  }
-  
-  .your-table-class th {
-    background-color: #f2f2f2;
-  }
