+++ web-dien-lanh/src/components/content/maintenance/Maintenance.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import './Maintenance.scss';
 import { useNavigate } from 'react-router-dom'
+import img1 from '../../../images/anhdv_vesinhmaylanh.jpg'
+import img2 from '../../../images/anhdv_vesinhmaygiat.jpg'
 const styles = {
    h4Style: {
       display: 'flex',
@@ -29,8 +31,8 @@ const Maintenance = () => {
 };
 
 const postData = [
-   { id: 1, imgUrl: 'https://hanara.vn/wp-content/uploads/2022/11/ve-sinh-may-giat-300x225.jpg', title: 'DỊCH VỤ BẢO TRÌ MÁY GIẶT - Giữ máy giặt luôn sạch đẹp, hoạt động hiệu quả', content: 'Quần áo khi giặt vẫn bị ám mùi khó chịu khi đã dùng nước xả',link: '/CWM' },
-   { id: 2, imgUrl: 'https://hanara.vn/wp-content/uploads/2022/10/dich-vu-bom-gas-may-lanh-1-300x225.jpg', title: 'Dịch Vụ BẢO TRÌ MÁY LẠNH', content: 'Qua quá trình sử dụng lâu dài kiến gas máy lạnh bị cạn kiệt, hết...',link: '/RACM' }
+   { id: 1, imgUrl: img2, title: 'DỊCH VỤ BẢO TRÌ MÁY GIẶT - Giữ máy giặt luôn sạch đẹp, hoạt động hiệu quả', content: 'Quần áo khi giặt vẫn bị ám mùi khó chịu khi đã dùng nước xả',link: '/cleaning-the-washing-machine' },
+   { id: 2, imgUrl: img1, title: 'Dịch Vụ BẢO TRÌ MÁY LẠNH', content: 'Qua quá trình sử dụng lâu dài kiến gas máy lạnh bị cạn kiệt, hết...',link: '/regular-air-conditioning-maintenance' }
 ];
 
 export default Maintenance;
