--- web-dien-lanh/src/components/content/maintenance/RACM/RACM.scss
@@ -1,10 +1,0 @@
-    // #wrapper {
-    //     #header {
-        
-    //         @media screen and (max-width:1000px) {
-    //             color: red;
-    //             font-size: 30px;
-    //         }
-    //     }
-
-    // }
